fix: add default mongodb uri and kill chrome only on prod

+++ Notifier/generators/color.generator.js
@@ -57,7 +57,7 @@ const closeBrowser = async () => {
   } catch {}
 
   // close zombie chrome processes
-  shell.exec('pkill chrome');
+  config.PRODUCTION && shell.exec('pkill chrome');
 };
 
 // remove branding colors (https://discord.com/branding)
